core: Share spec scaffolding across editions
The Library spec sandbox (tmpdir root, write_library, media entry builders)
becomes a shared context, ffprobe metadata fabrication becomes a shared
helper, and the export sandbox accepts full media-entry hashes — so the Pro
specs stop re-declaring all three.
